+++ src/components/Section.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Trash, Pin, Enter } from './SvgIcons';
 
 function Section(props) {
     function handleChange() {
@@ -16,9 +17,9 @@ function Section(props) {
     return (
         <div className="section">
             {props.editing ? <p onBlur={onEdit} suppressContentEditableWarning={true} contentEditable="true">{props.section}</p> : <p>{props.section}</p>}
-            <button className="themedButton" onClick={handleChange} style={{ left: '72%' }}>➡️</button>
-            {/* <button className="themedButton" onClick={this.onEdit}>✏️</button> */}
-            <button className="themedButton" onClick={onDelete} style={{ left: '87%' }}>❌</button>
+            <button className="themedButton" onClick={handleChange} style={{ left: '72%' }}><Enter></Enter></button>
+            {/* <button className="themedButton" onClick={onEdit}>✏️</button> */}
+            <button className="themedButton" onClick={onDelete} style={{ left: '87%' }}><Trash></Trash></button>
         </div>
     )
 }